docs(chronicle_compression): write custom configuration implementation for Chronicle Compression
