+++ src/sinks/gcp_chronicle/mod.rs
@@ -1,4 +1,5 @@
 #[cfg(feature = "sinks-gcp-chronicle")]
 pub mod chronicle_unstructured;
+pub mod compression;
 pub mod partitioner;
 pub mod sink;
